backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
openai==1.51.0
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
numpy==1.26.4
